moved group loading out of the action plugin
now action is solely responsible for updating the groups when
the time comes. addition of the cached groups to an authenticated
user will be done in the passthrough auth module (included).

The passthrough module simply loads the group file and merges AD
groups early enough in the loading stage so that they are visible to
AJAX requests, etc